Export get_relation_web to new file
